Modularize JavaScript
