+++ index.html
@@ -55,6 +55,6 @@
             </form>
         </div>
     </div>
-    <script src="script.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
